+++ sandbox/lib/navigation/navigation_panel.dart
@@ -4,9 +4,7 @@ import 'package:widgetbook_annotation/widgetbook_annotation.dart';
 
 import 'root_data.dart';
 
-@UseCase(name: 'Default', type: NavigationPanel)
+@UseCase(name: 'Default', type: NavigationPanel, cloudExclude: true)
 Widget navigationPanelDefaultUseCase(BuildContext context) {
-  return NavigationPanel(
-    root: root,
-  );
+  return NavigationPanel(root: root);
 }